+++ Changelog.md
@@ -13,6 +13,7 @@
 - Fixed bug preventing graders from creating new notes in results view (#4668)
 - Fixed bug preventing new tags from being created from results view (#4669)
 - Remove deprecated "detailed CSV" download link from submissions/browse (#4675)
+- Introduces Deductive Annotations (#4693)
 - Introduces annotation usage details panel to Annotations tab in admin settings (#4695)
 - Fixed bug where bonuses and deductions were not displayed properly (#4699)
 - Fixed bug where image annotations did not stay fixed relative to the image (#4706)